+++ libretroshare/src/serialiser/rstlvkeys.cc
@@ -471,7 +471,7 @@ bool  RsTlvKeySignature::GetTlv(void *data, uint32_t size, uint32_t *offset)
 		return false;	
 	
 	uint16_t tlvtype = GetTlvType( &(((uint8_t *) data)[*offset])  );
-	uint16_t tlvsize = GetTlvSize( &(((uint8_t *) data)[*offset])  );
+	uint32_t tlvsize = GetTlvSize( &(((uint8_t *) data)[*offset])  );
 	uint32_t tlvend = *offset + tlvsize;
 
 	if (size < tlvend)    /* check size */
